Move storage model
